Incomplete: An attempt at making a Generic Factory
This is not a complete solution. This was an attempt to use a
templated generic factory that would match a call to make
an instance to the templated constructor. This mechanism would
choose the most suitable constructor if it is overloaded, or fail
at compile time if no suitable constructor is found.

The most obvious code smell is that in order to get this to compile,
a dummy "IFactory<IBatchOperation>" class needs to be inherited from.
This is so that we do not need to declare the implementation of
BatchOperation in the submit_operations() signature. IFactory is
empty and does not actually contain the functions we are interested
in. Those instead lie with the GenericFactory which can't be listed
in the signature as GenericFactory needs to know the implementation.

I am leaving this here in case anyone wants to take a peek at the diff
and has an idea if this is worthwhile to pursue.