Use a friend class to gain access to Context's private attrs
Create a friend class of BatchContext to be able to peer into
BatchContext's private members. This class should not be used in any
production code. It is used to be able to seed the BatchContext with
mocked Operation data. This will be useful in the future if we
want to unit test what happens with completed/failed operations.