Add Factory mechanism to choose to make Real or Mock Operations
This allows for unit testing to switch what kind of IBatchOperation
gets made.

This also explores providing a testing queue to allow a test to
create and modify the mocks that get generated from the factory
for additional control.

Some changes here should not be included in the final PR. These
have been clearly marked, and are used to verify that the factory
is working as intended.